+++ packages/apolloschurchapp/src/ui/Onboarding/slides/AboutYou/AboutYou.stories.js
@@ -1,16 +1,11 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react-native';
+import { GradientOverlayImage } from '@apollosproject/ui-kit';
 
 import AboutYou from '.';
 
 storiesOf('Onboarding/slides/AboutYou', module)
   .add('default', () => <AboutYou />)
-  .add('withImg', () => (
-    <AboutYou
-      imgSrc={{ uri: 'https://picsum.photos/750/750?random' }}
-      setFieldValue={() => {}}
-    />
-  ))
   .add('slideTitle', () => (
     <AboutYou slideTitle={'Custom title text'} setFieldValue={() => {}} />
   ))
@@ -28,6 +23,16 @@ storiesOf('Onboarding/slides/AboutYou', module)
   ))
   .add('values – gender', () => (
     <AboutYou values={{ gender: 'Male' }} setFieldValue={() => {}} />
+  ))
+  .add('BackgroundComponent', () => (
+    <AboutYou
+      BackgroundComponent={
+        <GradientOverlayImage
+          source={'https://picsum.photos/640/640/?random'}
+        />
+      }
+      setFieldValue={() => {}}
+    />
   ))
   .add('Slide props', () => (
     <AboutYou onPressPrimary={() => {}} setFieldValue={() => {}} />